requirements.txt
# ============================================
# 러너웨이 백엔드 필수 패키지
# ============================================
# 설치 명령어: pip install -r requirements.txt
# ============================================

# --------------------------------------------
# 웹 프레임워크
# --------------------------------------------
# FastAPI - 현대적인 Python 웹 프레임워크 (gradio 6.1.0 호환)
fastapi>=0.115.2,<1.0

# Uvicorn - ASGI 서버 (FastAPI 실행에 필요)
uvicorn[standard]>=0.30.0

# --------------------------------------------
# 데이터베이스 관련
# --------------------------------------------
# SQLAlchemy - Python ORM (데이터베이스 작업을 쉽게)
sqlalchemy>=2.0.25

# PyMySQL - MariaDB/MySQL 연결 드라이버
pymysql>=1.1.0

# cryptography - PyMySQL SSL 연결에 필요
cryptography>=42.0.0

# --------------------------------------------
# 인증/보안 관련
# --------------------------------------------
# python-jose - JWT 토큰 생성 및 검증
python-jose[cryptography]==3.3.0

# passlib - 비밀번호 해싱
passlib[bcrypt]==1.7.4

# bcrypt - 비밀번호 암호화 알고리즘
bcrypt==4.1.2

# --------------------------------------------
# 환경 설정 관련
# --------------------------------------------
# python-dotenv - .env 파일에서 환경 변수 로드
python-dotenv>=1.0.0

# pydantic-settings - Pydantic 기반 설정 관리
pydantic-settings>=2.1.0

# --------------------------------------------
# HTTP 클라이언트 (카카오 API 호출용)
# --------------------------------------------
# httpx - 비동기 HTTP 클라이언트
httpx>=0.26.0

# starlette - FastAPI의 기반 프레임워크 (gradio 6.1.0 호환)
starlette>=0.40.0,<1.0

# --------------------------------------------
# 유틸리티
# --------------------------------------------
# python-multipart - 파일 업로드 처리 (gradio 6.1.0 호환)
python-multipart>=0.0.18

# email-validator - 이메일 유효성 검사
email-validator==2.1.0

# --------------------------------------------
# 개발 도구 (선택사항)
# --------------------------------------------
# pytest - 테스트 프레임워크
pytest==8.0.0

# pytest-asyncio - 비동기 테스트 지원
pytest-asyncio==0.23.3

# httpx - 테스트용 HTTP 클라이언트 (위에서 이미 설치)
